refactor(build): vercel serve and constants
