+++ packages/waku/src/lib/constants.ts
@@ -1,0 +1,3 @@
+export const EXTENSIONS = ['.js', '.ts', '.tsx', '.jsx', '.mjs', '.cjs'];
+export const SRC_MAIN = 'main';
+export const SRC_ENTRIES = 'entries';
