+++ packages/waku/src/lib/plugins/vite-plugin-rsc-analyze.ts
@@ -1,7 +1,7 @@
 import type { Plugin } from 'vite';
 import * as swc from '@swc/core';
 
-import { EXTENSIONS } from '../config.js';
+import { EXTENSIONS } from '../constants.js';
 import { extname } from '../utils/path.js';
 import { parseOpts } from '../utils/swc.js';
 // HACK: Is it common to depend on another plugin like this?
