+++ packages/waku/src/lib/plugins/vite-plugin-rsc-transform.ts
@@ -1,7 +1,7 @@
 import type { Plugin } from 'vite';
 import * as swc from '@swc/core';
 
-import { EXTENSIONS } from '../config.js';
+import { EXTENSIONS } from '../constants.js';
 import { extname } from '../utils/path.js';
 import { parseOpts } from '../utils/swc.js';
